+++ 07_web_endpoints/fasthtml_checkboxes.css
